+++ Changelog.md
@@ -1,6 +1,10 @@
 CHANGE LOG
 =========
-##Version 1.1 (Current)
+Version 1.5 (Current)
+  * Bulk Add Update: Add multiple items to the Database with the "Bulk Add" table
+  * You can now add the Standard/List prices when adding items
+
+Version 1.1
   * Added Change Log
   * Removed select button - You can now select and edit items by just checking their bubble
   * The user is now prevented from adding duplicate items with the same RM Number